make loading of RHD/LHD rules more robust against duplicates or conflicts
